Fix spacing issues with the subreddit manager and pinned subreddit bar
This removes a workaround for a bug in previous versions where having
both the subreddit manager enabled and the subreddit bar pinned caused
the subreddit bar to shrink to ten columns wide and three rows tall.

Since the bug seems to have dissapeared, the workaround was causing the
spacer to be too short.