Implement more aggressive caching of some XHR requests.
In some subreddits (`r/mylittlepony` is a great example) each page of
results can contain 12+ links that require an `API` request to inline.
While built in the browser cache does mitigate the issue some sites
(deviantart sets a `max-age` of only 3 minutes and tumblr doesn't seem
to request any caching) are unlikely to hit the cache much. If
`agressiveCache=true` is added to the background page request then the
successful XHR results will be stored in a simple in-memory cache.
Because greasemonkey does not provide a background page, a shared cache
can not be as easily or as cleanly implemented.

Default capacity is 250 entries.

Cache pruning algorithm is: whenever the `capacity` is reached, remove
the lowest weighted `capacity/2` entries from the cache based the
equation `weight=hits/age`

Aggressive caching is currently only enabled for DeviantART and tumblr.
Imgur is expected to generally appear more often, but it has a
comfortable `max-age` of 2 hours so it does not haveaggressivee caching
enabled.