Cleanup the per browser copies of the main js file.
First I removed the Safari copy of `reddit_enhancement_suite.user.js`
since none of the other browsers had it and because it was causing
frustration with branches.

Secondly I added the per browser `reddit_enhancement_suite.user.js`
files to `.gitignore` to keep the extra copies out of the git repo.