Attempt to preserve the original urls when following links.
I created a system for preserving the original link (when necessary)
without the trouble of loading entire webpages to mark history.

When necessary, the original link is saved to a property called
hrefOverride. If that is detected during expand creation then
mouseover/mouseout handlers will be attached to swap the href on demand.

This works well but it can cause the link to turn blue on mouseover.
Without marking the history this has been made nearly impossible to
prevent in the name of user privacy.