Fix message passing in Chrome
The recent update to Chrome 21 depreciated the `sendRequest` and
`onRequest.addListener` methods and replaced them with the seemingly
identical `sendMessage` and `onMessage.addListener` methods.

This broke a lot of things in Chrome, especially things that depend in
XHR. This may warrant an emergency release of the new version to
restore functionality.