Fixes issue #244
The incorrect date stamp was being used. The actual UTC stamp should be
used. This bug was reasonably niche in that it would only appear on the
boundary between two months for a new user created on that day. The
incorrect time zone caused the user to be created in the 'future',
causing a negative difference. Also, the month logic check was incorrect
so even the negative date was wrong, this has been fixed.

I also fixed the leapyear check which was incorrect and had the
potential for bugs at a later date.

I also fixed this for Subreddit Info which had the same issue