RESUtils.indexNestedArray, RESUtils.indexOptionTable
`.indexNestedArray(...)`, given an array of objects (or arrays), will:

* create a 1:\* key:item, unique-key index for the array's elements on the value of the element's keyIndex/keyFieldIndex property
* return an accessor function which, given a key, returns an item (element) via the index

Key comparison is case-insensitive.

Many keys can map to one item, particularly if the item's key property is a list delimited by `keyValueSeparator`.  `.indexOptionTable(...)` automatically sets this if the key field is [type=list].

If several items map to one key, the index maps the key to the last item.  To hell with many-to-many, although I suppose it would be easy enough to add a "manyToMany" parameter and, when setting the key-index mapping (`index[key] = item`), instead do `index[key].push(item)` if manyToMany is set.  But YAGNI.

---

`.indexNestedArray(...)` is a convenience wrapper around `.indexNestedArray()` for *module option[type=table] value* arrays.