Fix for HoverZoom/Character counter/BigEditor
Hover Zoom removes the title attribute on hover which was causing the
page to move up and down when hovering image links. Since the character
counter relied on that attribute for its CSS, it caused it to break
position and make the page move up/down by x amount as it lost/regained
its styling.

Big editor button styling was broken on wiki pages due to a bum selector
by yours truly.