Fixed some mostly harmless JS errors while fading the hover.
Previously hover.state was deleted as soon as the hover started to close. This meant that if the user moved the mouse back into the popup during the fade animation, some code that tried to access hover.state would trigger an error. The error does not seem to have a significant negative effect.

Signed-off-by: Scott McClaugherty <gamefreak4321@gmail.com>