Tell Instapaper to ignore key-navigation numbers
When clicking in a comment or in the text of a self post, RES decorates
each hyperlink with a superscripted number (like "[1]"). If the user
activates these numbers and then uses Instapaper's "Read Later"
bookmarklet to save the page's content for later viewing, these numbers
are included. This patch gives these numbers' elements the
"instapaper_ignore" class, making them invisible to Instapaper's parser.