addressed some timing issues with chrome related to async loading / The Great File Split of 2013/14...
essentially, it became possible to run code too early.

Already on the TODO list: really go back and revisit a lot of this architecture. It's sketchy because RES started as a GM script which used localStorage - a synchronous API... and rather than totally tear apart huge parts of RES, basically we're "faking" a synchronous API.

eventually we need to rewrite the core so that all modules assume storage requests are async.