Fix issues related to multireddits misbehaving
RES fails to handle multireddits correctly in at least two ways:

1. The hot/new/rising/etc tabs on a multireddit link to an incorrect
location if the multireddit is owned by a user other than the logged-in
user.
2. The commandline commands for switching to a different tab (`./t`,
`./c`, etc.) take you to an incorrect location if used on a multireddit
(owned by any user).

This commit addresses both those issues. The fix in utils.js is kind of
a kludge, but I don't want to risk breaking the `match` regex in
`RESUtils.currentUserProfile()`.