Add sort-by-gilded/promoted to commandline
In addition to the existing `/n`, `/t`, `/h`, and `/c` sort modes
already available from the RES commandline, this commit adds support for
`/g` (gilded) and `/p` (promoted/ads).

Note that you cannot view "promoted" content by a single user, and that
"sorting" by gilded for a single user is a bit different than sorting by
new/top/etc, so there is a bit of special handling for those cases.