Use RESUtils.profileRegex instead of a kludge
`RESUtils.profileRegex` seems to correctly identify multireddits as
non-user-profile pages, so we may as well use that rather than doing a
check for whether we're in a multireddit first.

`.currentUserProfile` was the only one of the four `.current*` functions
to have a regex hardcoded inside the function anyway, so this works out
well for future maintainability.