Added header embed option for subs w/ radd.it links in sidebar
If a subreddit has a radd.it link in their sidebar, a new option is added to the end of the hot/ top/ etc. <ul> labelled "radd.it player".  Clicking the new option adds an <iframe> to the top of the sidebar with an embedded version of the link.  Clicking it again removes the <iframe>.

Screenshot: http://i.imgur.com/bAgX2zX.png

Note: subreddits that do not have radd.it links in their sidebar are unaffected by this change.