Don't presume window is the global object for js
The bundled Favico lib exports itself by setting `this.Favico = Favico`.

The orangered module then presumes that Favico will be available on the
global `window` object, but that's not the case for Firefox's sandbox.
From [Bug 1208775][]'s discussion, it sounds like we might WONTFIX this.

Simply omitting `window.` works equally well in Chrome and Firefox.

[Bug 1208775]: https://bugzilla.mozilla.org/show_bug.cgi?id=1208775#c3