fix showImages on NER pages on /u/me/saved
The hack to use the second siteTable element (to account for sponsored links also being #siteTable.sitetable) caused the watchForElement siteTable observer to attach to the RES-saved siteTable instead of the reddit one, but it's no longer necessary.
This hack also happened to fix search pages (since the subreddit and post listings in search results are indistinguishable...), so we need to use `.search-result-listing:last` to avoid a regression there.