resolve a circular dependency between browser-specific `environment` impls and the default impl
By removing the _sendMessage and _addListener exports from the default impl, allowing them to be exported by the browsers before re-exporting the default impl, which depends on the browser-specific implementations of those functions.

To ensure that the lint still passes, make `chrome` the default build target for linting. This has the added benefit of making `npm run once`, etc. work without a BUILD_TARGET env var (defaulting to Chrome).