Shortened repetitive descriptions for ease of read
Current help menu is hard to read at a glance because it is tl;dr. I've trimmed descriptions of things that are just variants of their preceding item.

Another thing that needs to be done is to make the help menu obey night mode, but I'm not sure how to do that.