Setting link that can be relatively linked
Currently, since the current links begin with an exclamation mark (!) after the #, they cannot be relatively linked since reddit requires 'domains' to start with an alphanumeric character. So this is an invalid link:

`[RES Settings](#!settings)`

So I'm proposing changing the link syntax to:

`[RES Settings](#res/settings)`

So people on reddit can more easily link to reddit settings that work no matter what page (or page context level) you are viewing them from. 

This change is backwards compatible as well, so previous links will still work and be converted to the new format. The `hashComponents` doesn't have to be changed for this to work, but if it doesn't, it would greatly hinder the features discoverability.