Update Konami.js library (#3526)
* Add latest konami library to vendors folder

* Add exports statement to vendor Konami library

* Expose the Konami object through the vendor file

* Replace the Konami import and obtain it from vendor instead

* Remove old Konami dependency

* Leverage the Konami constructor and remove call to almostThere, dne

* Disable eslint for require'ing module not at the top level

* Update description, Enter is no longer required

* Modify original Konami.js to have 'almostThere'

'almostThere' is used to determine if the code has been almost entered,
this is useful because we can disable other shortcuts if we're about to
execute the Konami code.

* Revert all changes to original konami.js file

* Add useful properties to konami object and override load()

The original Konami code doesn't account for the code almost being
completed. If we check for this outside of the event handler, then race
conditions occur thus giving us incorrect behaviour. Calling `addEvent`
directly seemed like the best approach that worked.

* Delete konami vendor lib and port only used functionality

We don't need support for IE pre 11 or touch capable devices, so this
simplifies the code a lot. In addition, we needed to add support for
determining how far along the Konami code was.

* Remove Konami export from vendor/index.js, no longer exists

* cleanup some leftovers from old konami.js