Added descriptions to modules in settings console
The title attribute of the list items includes the complete module
description with HTML tags stripped. There is also a small text
block below the module title containing the first sentence of the
description

Modified method of parsing module description
Just use $ to fetch the text of a description in order to safely
handle possible HTML in the description. Per @andytuba request :smile:

Change how the module description is pushed to new line
switch from using a <br> tag to defining a CSS style with
`display: block`