improve account switcher error reporting
Previously, we assumed that if an error code could not be found in the message, login was successful. This fails when reddit changes their internal error codes (which happened).
Now, we check if `success` is true, and if we do not recognise the error, still alert the user instead of doing nothing.