Fixed referenced object on reused hover element
The hover menu for the subreddit manager gets reused after it's first creation, but doesn't update it's 'click' action references to the edit and delete button which tells it which subreddit to modify. So if you have more than 1 subreddit and try to edit one, it will always populate the information of the first object hovered.