Subreddit manager feature improvements
So this does two things, both of which I've been doing for a while with a hack and a stylesheet tweak. 

1. Option to highlight a multi on the subreddit manager bar if one of it's children are selected. So if I have a `Gaming` multi the target of `overwatch+leagueoflegends+dota2+heroesofthestorm`, if I'm on any of those subreddits, `Gaming` will get highlighted as if it were the current sub. For people that utilize the multi's mainly for the dropdown, this should make it more intuitive for them if they want to enable it. Also makes the next thing more useful. 

2. Allows for an alternative use of the dropdown to act as a normal, single subreddit with "related subs" in the dropdown. In use, it looks like this:

Subreddit: `leagueoflegends?+leagueoflegends/new?+leagueofmeta+loleventvods`
Display Name: `leagueoflegends ▼`
Results: https://gfycat.com/MenacingWastefulCorydorascatfish

The inclusion of the `?` on the first subreddit is all you need. This is how I made it work before because the URL would ignore everything after that point. Shouldn't negatively affect anything else and is pretty simplistic in how to make use of it.