newCommentCount: only use sharded storage for comment counts, not subscriptions
Sharded storage is good for comment counts, because:
a. keys are independent, so we can update different posts on different tabs without conflict
b. querying for one key is constant time, regardless of how many are stored
However, iterating through (/loading) all of them is very expensive (_everything_ has to be pulled out of storage), and is necessary to check for subscriptions.
So use a single object to hold all subscriptions.

This also has a minor improvement in that it doesn't overwrite the comment counts when showing a notification for a subscribed post, so if you click the notification and load the post you'll still see the number of new comments.