Remove hard-coded config for expando scrolling
This fixes a regression from:
https://github.com/honestbleeps/Reddit-Enhancement-Suite/commit/19028db28c5d24508ec647587426a484002237ab#diff-81e82c385dc9912064f4bc5e91b47e58L755

Before:
```
if (options.mediaBrowseScrollStyle) options.scrollStyle = options.mediaBrowseScrollStyle;
```

After:
```
options.scrollStyle = 'top';
```

Hard-coding config causes a regression of issue:
https://github.com/honestbleeps/Reddit-Enhancement-Suite/issues/2628

Issue summary:
Using `scrollStyle = 'top'` causes expanded items to scroll up and off the
top of the window when `on hide move down` is used.

Not sure why this was hard-coded. My guess: a desire to have a different
scroll style for navigating comments vs media. If so, then a new setting needs
to be introduced. Until then, stop hardcoding and use the settings.