remove headReady
Fun fact: stuff that normally goes in <head> actually works just fine outside of it, so we don't need to wait for it.
This allows some stages to start earlier.
And it allows us to remove a dependency on /core from inside /utils, which will prevent everything from exploding when adding/rearranging imports in init (or at least work towards that).