fix inference infinite recursion in dashboard
Basically, the T in Wrapper::get() and Wrapper::set() were dependant on each other in a cycle, since the loaded value couldn't be inferred from context, and the loaded value is just mutated and then stored.
Flow generally breaks cycles like this, so something else weird must have been going on to cause this.