No longer sending the full options object to background script when redirect module's state changes
This change also fixes a bug in firefox - sending the entire options object to background scripts
was resulting in a 'The object could not be cloned' error being thrown. Firefox uses the structured
cloning algorithm to copy objects send in messages. The full options object contains functions,
which cannot be handled by SCA, and on encountering which SCA will throw a DATA_CLONE_ERR exception