Optional permissions in Firefox (#4989)
Firefox requires that optional permissions requests are initiated from an included page (among a small set of other options).

This opens a popup which prompts the user to click a "request permissions" button in order for Firefox to open the actual prompt. This solution is certainly not ideal, but allows us to finally make a lot of permissions in Firefox optional.