showImages: Fix conserveMemory on Chromium 75
It appears that (a possible bug) in Chromium causes it to not register intersections when the element width / height is 0. When unloading images / videos, the expando media element width becomes 0, causing it to never again be registered as intersecting.

This checks intersection on the media parent element, which width isn't changed when unloading.