New module: autoHide for hiding seen posts
When the bottom (the buttons-row) of a post is shown for more than 2s continuously, the post is hidden. This may be tweaked in the future to reduce false negatives / false positives, e.g. by adjusting the IntersectionObserver margins / target, or improving the detection heuristic.

By default only comments are marked as read (via readComments), as comments are hidden by RES and can be easily undone. Link posts are hidden via thingHide, equivalent of clicking the 'hide' button. Changes are sent every 5 seconds, and flushed on `beforeload`.