jQuery as direct dependency (#5142)
* Remove dependency tokenInput

tokenInput is only used for auto-completing subreddits in the dashboard, and is not really that useful anymore since it's slow and not that helpful due the number of possible subreddits.

* jQuery as direct dependency

safeHtml is replaced DOMPurify.santize, to make its usage more transparent and less error-prone. The net result is about the same, as its default options appears to be sufficient for our purposes.

The guiders script is modified so that it export `guiders` directly.