Fix nightMode / subredditStyleToggle race condition
A change in how nightMode / subredditStyleToggle communicated with each other caused a race condition when nightMode is enabled and the subreddit has a nightMode-compatible theme.

When opening a new tab, the new state went through:
1. Turn off custom style, since it could not at that stage be determined that the subreddit supported it
2. Turn it back on when support is estabished

This state was broadcasted to every other tab. That was unnecessary, and is now disabled by using the new customToggle mode 'autoLocal'.