Relax regex on twitch clip names.
Twitch has a few additional formats for the clip names, including double
hyphens and also ending with a hyphen. Relaxed the Regex for twitch clip
name matching so that the hyphen character will be treated as one of the
characters in the randomly generated string.

This means this will now match any additional string appended that uses
the hyphen or the characters from the \w set (A-Za-z0-9_).