Fixed: userInfo popup not showing on crossposts
In crossposted posts the user tag was not visible and the user info popup didn't show when hovering. See screenshot: https://i.imgur.com/MlAPJFT.png

The cause was that the element with CSS-class "tagline" is a <p> in normal posts, but it's a <div> in crospossted elements. Removing the element type on usernameSelector fixes this.

Another option would be to keep "p.tagline a.author" and add "div.tagline a.author".